RLTypes

Documentation for RLTypes.

This package provides shared functionality for the reinforcement learning framework. The package is not only required by the specific algorithm packages, but also exports names into the main module of Julia. The package contains important types and functions that are shared between the packages:

DDPG

DQN

MBRL

NNDynamics

NODEDynamics

ODERNNDynamics

Rewards

The main purpose of the package is to provide

a type system for the reinforcement learning framework to enable multiple dispatch,

set hyperparameters automatically, or at least provide a common interface for setting hyperparameters,

to differentiate function signatures and hence enable multiple dispatch,

provide shared functionality, e.g. Replay Buffers.

The majority of the functionality is used internally, but certain function signatures set by the user depend on RLTypes.

Installation

In the julia REPL, run

using Pkg
Pkg.add(url="https://github.com/SvenDuve/RLTypes.jl")

Bring package into scope with

using RLTypes

Example usage

# to be used within the function call to an RL Algorithm to set hyperparameters
AgentParameter(training_episodes=100, batch_size=128)

Check the source code for a full list of parameters.

Function Reference

RLTypes.AgentParameter

RLTypes.DiscreteReplayBuffer

RLTypes.DiscreteReplayBuffer

RLTypes.EnvParameter

RLTypes.ModelParameter

RLTypes.ReplayBuffer

RLTypes.ReplayBuffer

RLTypes.collectTransitions!

RLTypes.remember

RLTypes.sample

RLTypes.sample

RLTypes.sample

RLTypes.sample

RLTypes.AgentParameter — Type

AgentParameter()

Agent Parameters to be set by the user at function call.

Example

AgentParameter(training_episodes=100)

RLTypes.DiscreteReplayBuffer — Type

DiscreteBuffer(;capacity, memory, pos)

A type for a discrete action space Buffer.

RLTypes.DiscreteReplayBuffer — Method

DiscreteReplayBuffer(capacity::Int)

Outer constructor for the discrete action space buffer.

Example

buffer = DiscreteReplayBuffer(1000000)

RLTypes.EnvParameter — Type

EnvParameter()

Parameters for the environment. This is called from inside the respective module.

RLTypes.ModelParameter — Type

ModelParameter()

Parameters to be set by the user for training the model of the environement.

Example:

ModelParameter(train=true, collect_train=100, collect_test=10, train_start=1000)

RLTypes.ReplayBuffer — Type

ReplayBuffer(;capacity, memory, pos)

A type for the buffer.

RLTypes.ReplayBuffer — Method

ReplayBuffer(capacity::Int)

Outer constructor for the buffer.

Example

buffer = ReplayBuffer(1000000)

RLTypes.collectTransitions! — Method

collectTransitions!(buffer, env, n_transitions)

Collects n_transitions from the environment (env) and stores them in the buffer.

Examples

collectTransitions!(buffer, env, 1000)

RLTypes.remember — Method

remember(buffer::Buffer, state, action, reward, next_state, done)

Stores tuples of transitions in the buffer.

Example

remember(buffer, s, a, r, s´, t)

RLTypes.sample — Method

sample(buffer::DiscreteReplayBuffer, method::AgentMethod, batch_size::Int)

Sample a batch of agent type (ie. in random order) transitions from the discrete action space buffer.

Example

S, A, R, S', T = sample(buffer, AgentMethod(), 64)

RLTypes.sample — Method

sample(buffer::DiscreteReplayBuffer, method::ModelMethod, batch_size::Int)

Sample a batch of model type (ie. in sequential order) transitions from buffer.

Example

S, A, R, S', T = sample(buffer, ModelMethod(), 64)

RLTypes.sample — Method

sample(buffer::Buffer, method::AgentMethod, batch_size::Int)

Samples a batch of agent type (ie. in random order) transitions from the buffer.

Example

S, A, R, S', T = sample(buffer, AgentMethod(), 64)

RLTypes.sample — Method

sample(buffer::ReplayBuffer, method::ModelMethod, batch_size::Int)

Sample a batch of model type (ie. in sequential order) transitions from buffer.

Example

S, A, R, S', T = sample(buffer, ModelMethod(), 64)